Add support for events
