+++ apidoc/ti_win_yaml.js
@@ -3,310 +3,279 @@
  * Licensed under the terms of the Apache Public License.
  * Please see the LICENSE included with this distribution for details.
  */
-var lbl = require('line-by-line'),
-    path = require('path'),
-	fs = require('fs.extra'),
-	TI_KIT_SRC_DIRECTORY = '../Source/TitaniumKit/src',
-	TI_WIN_DOC_DIRECTORY = '../apidoc/Titanium',
-	TI_WIN_SRC_DIRECTORY = '../Source/';
+var path = require('path'),
+	fs   = require('fs.extra'),
+	_    = require('underscore'),
+	LineReader = require('line-by-line'),
+	titaniumKit = {},
+	windowsTitaniumKit = {},         // Stores what's implemented in Titanium Windows
+	windowsTitaniumKit_Missing = []; // Stores missing module names
 
-// Add 'contains' prototype to String
-if (typeof String.prototype.contains === 'undefined') {
-	String.prototype.contains = function(str) {
-		return this.indexOf(str) != -1;
-	}
-}
+String.prototype.capitalize = function() {
+    return this.charAt(0).toUpperCase() + this.slice(1);
+};
 
-// Define signatures
-var property_ro_signature = 'TITANIUM_ADD_PROPERTY_READONLY(';
-var property_signature = 'TITANIUM_ADD_PROPERTY('
-var method_signature = 'TITANIUM_ADD_FUNCTION(';
-var namespace_signature = 'namespace ';
-
-var properties_found = 0;
-var methods_found = 0;
-var events_found = 0;
-
-var docs = {};
-
-// Walk through the TitaniumKit source directory
-var walker = fs.walk(TI_KIT_SRC_DIRECTORY)
-walker.on("file",
-	function (root, stat, next) {
-		if (stat.name.contains('.cpp')) {
-		    processClass(root, stat.name, next);
-		} else {
-			next();
-		}
-	}
-);
-walker.on("end", processEvents);
-
-function processEvents () {
-	var recurse = fs.walk(TI_WIN_SRC_DIRECTORY);
-	recurse.on("file", function (root, stat, next) {
-		if (stat.name.contains('.cpp')) {
-		    processEvent(root, stat.name, next);
-		} else {
-			next();
-		}
-	});
-	recurse.on('end', function() {
-		var className, cls;
-		for (className in docs) {
-			cls = docs[className];
-			generateYAML(cls.root, cls.module_name, cls.module, cls.properties, cls.methods, cls.events);
-		}
-		console.log('\nProcessing whitelist.txt...');
-		processWhitelist('whitelist.txt', function() {
-			console.log('\nFound '+methods_found+' methods');
-			console.log('Found '+properties_found+' propeties');
-			console.log('Found '+events_found+' events');
-			console.log('Total APIs : '+(methods_found+properties_found+events_found));
-			
-			var docPath = process.argv[1].replace(/\\/g, '/')
-			docPath = docPath.substring(0, docPath.lastIndexOf('/'))+'/Titanium';
-			console.log('\nTo append the documentation run the following command from titanium_mobile/apidoc :\n\nnode docgen.js -f parity -a '+docPath);
-		});
-	});
-}
-
-function processEvent(root, file, next) {
-	var l = new lbl(path.join(root, file)),
-		name = null,
-		events = [],
-		namespace = [],
-		module = file.replace('.cpp', '').replace('Module', '');
-		if (module === 'TitaniumWindow') {
-			module = 'App';
-		}
-	l.on('line', function(line) {
-		if (~line.indexOf('fireEvent') && !~line.indexOf('//') && !~line.indexOf('%') && !~line.indexOf('TITANIUM_LOG')) {
-			if ((name = line.split('"')[1]) && !~events.indexOf(name)) {
-				events.push(name);
-			}
-		} else if (~line.indexOf(namespace_signature) && !~line.indexOf('//') && !~line.indexOf('using')) {
-			var ns = line.trim().split(' ')[1].replace('Module','');
-			switch (ns) {
-				case 'TitaniumWindows':
-					ns = 'Titanium';
-					break;
-				case 'GlobalObject':
-					ns = 'Global';
-					namespace = [];
-					break;
-				case 'WindowsXaml':
-					ns = 'Windows';
-					break;
-			}
-			namespace.push(ns);
-		}
-	});
-	l.on('error', function (err) {
-		console.log('Error : ' + err);
-	});
-	l.on('end', function () {
-		var ns,
-			black_list = ['WindowsViewLayoutDelegate', 'TitaniumWindows'];
-		namespace.push(module);
-		ns = namespace.join('.');
-		if (namespace[0] === 'Titanium' && !~namespace.indexOf('Windows') && !~black_list.indexOf(module) && events.length > 0) {
-			events_found += events.length;
-			if (ns in docs) {
-				docs[ns].events = docs[ns].events.concat(events);
+//
+// Extract TitaniumKit API and find Windows implementation
+//
+function extract(callback) {
+	var walker = fs.walk(path.join(__dirname, '../Source/TitaniumKit/src'));
+	walker.on('file',
+		function (root, stat, next) {
+			if (/\.cpp$/.test(stat.name)) {
+			    extractTitaniumKit(root, stat.name, next);
 			} else {
-				docs[ns] = {
-					root: root,
-					module_name: ns,
-					module: module,
-					properties: [],
-					methods: [],
-					events: events
-				}
+				next();
 			}
+		}
+	);
+	walker.on('end', function() {
+		extractWhiteList(callback);
+	});
+}
+
+//
+// Extract TitaniumKit API from cpp source
+//
+function extractTitaniumKit(root, file, next) {
+	var reader = new LineReader(path.join(root, file)),
+		module_name, properties = [] , methods = [], events = [];
+	reader.on('line', function(line) {
+		var namespace_match = line.match(/\s*\:\s*\w+\(js_context,\s*\"([\w\.]+)\"\)/);
+		var unknown_match   = line.match(/\s*\:\s*Module\(js_context,\s*apiName\)/);
+		var property_match  = line.match(/TITANIUM_ADD_PROPERTY\w*\((\w+),\s*(\w+)\);/);
+		var function_match  = line.match(/TITANIUM_ADD_FUNCTION\((\w+),\s*(\w+)\);/);
+
+		if (namespace_match) {
+			module_name = namespace_match[1];
+		} else if (unknown_match) {
+			//
+			// Some modules are using special pattern for now...
+			//
+			if (file == 'IOStream.cpp') {
+				module_name = 'Titanium.IOStream';
+			} else if (file == 'View.cpp') {
+				module_name = 'Titanium.UI.View';
+			}
+		}
+
+		// Don't expose internal property/function which starts with "_"
+		if (property_match && property_match[2][0] != '_') {
+			properties.push(property_match[2]);
+		}
+		if (function_match && function_match[2][0] != '_') {
+			methods.push(function_match[2]);
+		}
+	});
+	reader.on('error', function(err) {
+		console.error(err);
+		process.exit(1);
+	});
+	reader.on('end', function() {
+		if (module_name) {
+			titaniumKit[module_name] = {
+				module_name: module_name,
+				properties: properties,
+				methods: methods,
+				events: events
+			};
 		}
 		next();
 	});
 }
 
-
-// Extract titanium properties and methods from the class file
-function processClass(root, file, next) {
-
-	var module_name = null;
-	var namespaces = [];
-	var module = null;
-	var properties = [];
-	var methods = [];
-
-    var l = new lbl(path.join(root, file));
-	l.on('line',
-		function (line) {
-			
-			var isProperty = line.contains(property_signature);
-			var isPropertyRo = line.contains(property_ro_signature);
-			var isMethod = line.contains(method_signature);
-
-			if (isProperty || isPropertyRo || isMethod) {
-				if (module == null) {
-					if (isProperty) {
-						module = line.substr(line.indexOf(property_signature)+property_signature.length).split(',')[0].replace('Module','');
-					} else if (isPropertyRo) {
-						module = line.substr(line.indexOf(property_ro_signature)+property_ro_signature.length).split(',')[0].replace('Module','');
-					} else if (isMethod) {
-						module = line.substr(line.indexOf(method_signature)+method_signature.length).split(',')[0].replace('Module','');
-					}
-				}
-
-				var name = line.split(',')[1].split(')')[0].substring(1);
-
-				if (module_name == null) module_name = namespaces.join('.')+(module != '' ? '.'+module : '');
-				if (module_name == 'Titanium.GlobalObject') module_name = 'Global';
-				if (module_name == 'Titanium' && module == '') module_name = 'Titanium.Proxy';
-				if (module_name == 'Titanium.Ti') module_name = 'Titanium';
-				if (module_name == 'Titanium.UI.TwoDMatrix') module_name = 'Titanium.UI.2DMatrix';
-
-				if (isProperty || isPropertyRo) properties.push(name);
-				if (isMethod && name[0] != '_') methods.push(name);
-
-			} else if (line.contains(namespace_signature)) {
-				var namespace = line.substr(line.indexOf(namespace_signature)+namespace_signature.length).replace('Module','');
-				if (namespace.contains(' ')) namespace = namespace.split(' ')[0];
-				if (namespace.contains('{')) namespace = namespace.split('{')[0];
-				if (!namespace.contains('.') && !namespace.contains(':') && !namespace.contains(';'))
-					namespaces.push(namespace);
-			}
-		}
-	);
-	l.on('error',
-		function (err) {
-			console.log('Error : ' + err);
-		}
-	);
-	l.on('end',
-		function () {
-			properties_found += properties.length;
-			methods_found += methods.length;
-			if (module != null) {
-				if (module == '') module = 'Titanium';
-				docs[module_name] = {
-					root: root,
+//
+// Load API from whitelist, because some API are implemented in JavaScript
+//
+function extractWhiteList(callback) {
+	var titaniumKit_whitelist = {};
+	var reader = new LineReader(path.join(__dirname, 'whitelist.txt')),
+		module_name, properties = [], methods = [], events = [];
+	reader.on('line', function(line) {
+		var element = line.split(' '),
+			type = element[0];
+		if (type === 'C') {
+			module_name = element[1];
+		} else if (type === 'P') {
+			properties.push(element[1]);
+		} else if (type === 'M') {
+			methods.push(element[1]);
+		} else if (type === 'E') {
+			events.push(element[1]);
+		} else if (line.trim() == "") {
+			// there must be a empty line which indicates separator of data
+			if (module_name) {
+				titaniumKit_whitelist[module_name] = {
 					module_name: module_name,
-					module: module,
-					properties: properties || [],
-					methods: methods || [],
-					events: []
-				}
+					properties: properties,
+					methods: methods,
+					events: events
+				};
 			}
-			next();
+
+			module_name = null;
+			properties = [], methods = [], events = [];
 		}
-	);
-}
-
-// Generate a YAML file from the extracted information
-function generateYAML(root, module_name, module, properties, methods, events, next) {
-	var folder = root.split('/').pop();
-
-	var yaml = 'name: '+module_name+'\n'+
-	           'platforms: [windowsphone]\n\n';
-	if (properties.length > 0) yaml += 'properties:\n';
-	for(i in properties) {
-		var property = properties[i];
-	    yaml += '  - name: '+property+'\n'+
-	            '    platforms: [windowsphone]\n';
-	}
-	if (methods.length > 0) yaml += 'methods:\n';
-	for (i in methods) {
-		var method = methods[i];
-		yaml += '  - name: '+method+'\n'+
-	            '    platforms: [windowsphone]\n';
-	}
-	if (events.length > 0) yaml += 'events:\n';
-	for (i in events) {
-		var event = events[i];
-		yaml += '  - name: '+event+'\n'+
-	            '    platforms: [windowsphone]\n';
-	}
-
-	var target = TI_WIN_DOC_DIRECTORY+(folder != 'src' ? '/'+folder+'/' : '/');
-
-	fs.mkdirp(target,
-		function (err) {
-			if (err) {
-				console.log(err);
+	});
+	reader.on('error', function(err) {
+		console.error(err);
+		process.exit(1);
+	});
+	reader.on('end', function() {
+		// merge the result
+		for (name in titaniumKit_whitelist) {
+			if (titaniumKit[name]) {
+				titaniumKit[name].properties = _.union(titaniumKit[name].properties, titaniumKit_whitelist[name].properties);
+				titaniumKit[name].methods    = _.union(titaniumKit[name].methods, titaniumKit_whitelist[name].methods);
+				titaniumKit[name].events     = _.union(titaniumKit[name].events, titaniumKit_whitelist[name].events);
 			} else {
-				fs.writeFile(target+module+'.yml', yaml,
-					function(err) {
-					    if(err) {
-					        console.log(err);
-					    } else {
-					        console.log('Generated '+module+'.yml');
-					    	if (next) next();
-					    }
+				titaniumKit[name] = titaniumKit_whitelist[name];
+			}
+		}
+
+		extractWindowsKit(callback);
+	});
+}
+
+//
+// Search for platform implementation and find what's implemented
+// 
+function extractWindowsKit(callback) {
+	var walker = fs.walk(path.join(__dirname, '../Source/'));
+	walker.on('file',
+		function (root, stat, next) {
+			if (/\.hpp$/.test(stat.name)) {
+				var reader = new LineReader(path.join(root, stat.name)),
+					module_name, 
+					properties = [], methods = [], events = [],
+					un_properties = [], un_methods = [], un_events = [];
+				reader.on('line', function(line) {
+					var module_match = line.match(/\s*TITANIUM_MODULE_UNIMPLEMENTED\(\"*(\w[\w.]*)\"*\)/);
+					if (module_match) {
+						windowsTitaniumKit_Missing.push(module_match[1]);
 					}
-				);
+
+					// C++ header should contain @ingroup to represent Titanium proxy name
+					var class_match = line.match(/@ingroup\s+([\w.]+)/);
+					if (class_match) {
+						module_name = class_match[1];
+					}
+					// Unimplemented Ti method be marked with TITANIUM_FUNCTION_UNIMPLEMENTED
+					var method_match = line.match(/\s*TITANIUM_FUNCTION_UNIMPLEMENTED\(\"*(\w+)\"*\)/);
+					if (method_match) {
+						un_methods.push(method_match[1]);
+					}
+					// Unimplemented Ti property be marked with TITANIUM_PROPERTY_UNIMPLEMENTED
+					var property_match = line.match(/\s*TITANIUM_PROPERTY_UNIMPLEMENTED\(\"*(\w+)\"*\)/);
+					if (property_match) {
+						var property_name   = property_match[1];
+						var property_name_C = property_name.capitalize();
+						un_properties.push(property_name);
+						// Disable property getter/setter too
+						un_methods.push('get' + property_name_C);
+						un_methods.push('set' + property_name_C);
+					}
+
+				});
+				reader.on('error', function(err) {
+					console.error(err);
+					process.exit(1);
+				});
+				reader.on('end', function() {
+					if (module_name && titaniumKit[module_name]) {
+						windowsTitaniumKit[module_name] = {
+							module_name: module_name,
+							properties: _.difference(titaniumKit[module_name].properties, un_properties),
+							methods: _.difference(titaniumKit[module_name].methods, un_methods),
+							events: _.difference(titaniumKit[module_name].events, un_events)
+						};
+					}
+					next();
+				});
+			} else {
+				next();
 			}
 		}
 	);
+	walker.on('end', callback);
 }
 
-// Generate whitelist YAML files
-function processWhitelist(file, next) {
-	var module_name = null;
-	var namespaces = [];
-	var module = null;
-	var properties = [];
-	var methods = [];
-	var events = [];
+//
+// Export API document to YAML
+//
+function exportYAML() {
+	for (module_name in titaniumKit) {
+		var lines = [];
+		lines.push('name: ' + module_name);
+		lines.push('platforms: [windowsphone]');
 
-    var l = new lbl(file);
-	l.on('line',
-		function (line) {
-
-			var element = line.split(' '),
-				type = element[0];
-
-			// Class
-			if (type === 'C') {
-				if (module != null) {
-					var folder = module_name.substring(module_name.split('.')[0].length+1, module_name.length-module.length-1).replace(/\./g, '/');
-					generateYAML(folder, module_name, module+'_white', properties, methods, events);
-					module_name = null;
-					namespaces = [];
-					module = null;
-					properties = [];
-					methods = [];
-					events = [];
-				}
-				module_name = element[1];
-				var split_module = module_name.split('.');
-				module = split_module[split_module.length-1];
-			
-			// Property
-			} else if (type === 'P') {
-				var name = element[1];
-				properties.push(name);
-
-			// Method
-			} else if (type === 'M') {
-				var name = element[1];
-				methods.push(name);
-
-			// Event
-			} else if (type === 'E') {
-				events.push(element[1]);
-			}
+		// dump properties
+		titaniumKit[module_name].properties.length > 0 && lines.push('properties:');
+		for (var i = 0; i < titaniumKit[module_name].properties.length; i++) {
+			var property  = titaniumKit[module_name].properties[i];
+			lines.push('  - name: ' + property);
+			lines.push('    platforms: [windowsphone]');
 		}
-	);
-	l.on('error',
-		function (err) {
-			console.log('Error : ' + err);
+
+		// dump methods
+		titaniumKit[module_name].methods.length > 0 && lines.push('methods:');
+		for (var i = 0; i < titaniumKit[module_name].methods.length; i++) {
+			var method    = titaniumKit[module_name].methods[i];
+			lines.push('  - name: ' + method);
+			lines.push('    platforms: [windowsphone]');
 		}
-	);
-	l.on('end',
-		function () {
-			var folder = module_name.substring(module_name.split('.')[0].length+1, module_name.length-module.length-1).replace(/\./g, '/');
-			generateYAML(folder, module_name, module+'_white', properties, methods, events, function() {next();});
+
+		// dump events
+		titaniumKit[module_name].events.length > 0 && lines.push('events:');
+		for (var i = 0; i < titaniumKit[module_name].events.length; i++) {
+			var eventname = titaniumKit[module_name].events[i];
+			lines.push('  - name: ' + eventname);
+			lines.push('    platforms: [windowsphone]');
 		}
-	);
+
+		// newline at end of file
+		lines.push('\n');
+
+		// dump to file
+		var modulepath = module_name.split('.');
+		var classname  = modulepath[modulepath.length-1];
+
+		modulepath.pop();
+		// everything should be go to Titanium directory
+		if (modulepath[0] != 'Titanium') {
+			modulepath[0] = 'Titanium';
+		}
+		var outdir = path.join(modulepath.join(path.sep));	
+		var outfile = path.join(outdir, classname + '.yml');
+		console.log('Generating ' + outfile);
+		fs.mkdirpSync(path.join(__dirname, outdir));
+		fs.writeFileSync(path.join(__dirname, outfile), lines.join('\n'));
+	}
 }
+
+function isEmpty(prop) {
+	return prop.properties.length == 0 && prop.methods.length == 0 && prop.events.length == 0;
+}
+//
+// Extract API from source & export to YAML
+//
+extract(function() {
+	// remove unimplemented API
+	for (var i = 0; i < windowsTitaniumKit_Missing.length; i++) {
+		delete titaniumKit[windowsTitaniumKit_Missing[i]];
+	}
+
+	// now merge the supported API
+	for (name in titaniumKit) {
+		var winprop = windowsTitaniumKit[name];
+		if (winprop) {
+			titaniumKit[name] = winprop;
+		}
+		if (isEmpty(titaniumKit[name])) {
+			delete titaniumKit[name];
+		}
+	}
+
+	exportYAML();
+});
+
